refactor: split olf.openmetadata into config/payloads/reconciliation/orchestration
Decompose the 724-line flat olf/openmetadata.py into a package split
by capability, per issue #101: _config (deploy configuration and
descriptor source resolution), _payloads (domain/data-product payload
construction), _reconciliation (REST upsert/resolve primitives via a
new OpenMetadataReconciler collaborator), and _orchestration (spec
building and deploy() sequencing on OpenMetadataDeployer, which
delegates to the reconciler). OpenMetadataDeployer keeps its original
method names as thin delegators so existing monkeypatch-based test
seams and the deploy() call sequence are unchanged. Tests reorganized
to mirror the new boundaries; no CLI behavior change.
