--- tools/olf/olf/openmetadata.py
@@ -1,724 +1,0 @@
-"""OpenMetadata governance metadata seeding over the REST API.
-
-Port of scripts/local/artifacts/openmetadata-metadata-deploy.sh (a bash wrapper
-around an embedded Python REST client). The seeding is imperative today; see
-docs/technical-debt.md for the reconciliation follow-up. Behavior is preserved
-exactly — only the host moved from a heredoc into this shared module.
-"""
-
-from __future__ import annotations
-
-import json
-import time
-import urllib.parse
-from collections.abc import Iterator
-from dataclasses import dataclass
-from pathlib import Path
-
-from openlakeforge_domain import load_domain_descriptor, load_domain_inventory_from_descriptors
-
-from olf.clients.openmetadata import OpenMetadataClient, OpenMetadataError
-
-__all__ = ["OpenMetadataClient", "OpenMetadataError", "OpenMetadataConfig", "OpenMetadataDeployer"]
-
-
-@dataclass
-class OpenMetadataConfig:
-    base_url: str
-    admin_email: str
-    admin_password: str
-    metadata_root: Path
-    metadata_source_dir: str
-    allow_missing_assets: bool
-    catalog_service: str
-    catalog_database: str
-    cleanup_legacy_default_database: bool
-    catalog_database_fqn: str
-    catalog_silver_schema_fqns: dict
-    catalog_gold_schema_fqns: dict
-    storage_service: str
-    storage_display_name: str
-    storage_endpoint: str
-    storage_region: str
-    storage_bronze_bucket: str
-    storage_silver_bucket: str
-    storage_gold_bucket: str
-
-    @classmethod
-    def from_environment(
-        cls,
-        environ,
-        *,
-        base_url: str,
-        admin_email: str,
-        admin_password: str,
-        metadata_root: str,
-        metadata_source_dir: str,
-        allow_missing_assets: bool,
-        catalog_service: str,
-        catalog_database: str,
-        cleanup_legacy_default_database: bool,
-    ) -> OpenMetadataConfig:
-        catalog_service = catalog_service or "polaris"
-        catalog_database = catalog_database or "lakehouse_dev"
-        catalog_database_fqn = environ.get(
-            "OPENLAKEFORGE_CATALOG_DATABASE_FQN", f"{catalog_service}.{catalog_database}"
-        )
-        silver_schema_fqns_raw = environ.get("OPENLAKEFORGE_CATALOG_SILVER_SCHEMA_FQNS_JSON")
-        gold_schema_fqns_raw = environ.get("OPENLAKEFORGE_CATALOG_GOLD_SCHEMA_FQNS_JSON")
-        return cls(
-            base_url=base_url.rstrip("/"),
-            admin_email=admin_email,
-            admin_password=admin_password,
-            metadata_root=Path(metadata_root),
-            metadata_source_dir=metadata_source_dir,
-            allow_missing_assets=allow_missing_assets,
-            catalog_service=catalog_service,
-            catalog_database=catalog_database,
-            cleanup_legacy_default_database=cleanup_legacy_default_database,
-            catalog_database_fqn=catalog_database_fqn,
-            catalog_silver_schema_fqns=(
-                _parse_json_env("OPENLAKEFORGE_CATALOG_SILVER_SCHEMA_FQNS_JSON", silver_schema_fqns_raw)
-                if silver_schema_fqns_raw
-                else _default_schema_fqns(metadata_root, metadata_source_dir, catalog_database_fqn, "silver")
-            ),
-            catalog_gold_schema_fqns=(
-                _parse_json_env("OPENLAKEFORGE_CATALOG_GOLD_SCHEMA_FQNS_JSON", gold_schema_fqns_raw)
-                if gold_schema_fqns_raw
-                else _default_schema_fqns(metadata_root, metadata_source_dir, catalog_database_fqn, "gold")
-            ),
-            storage_service=environ.get("OPENLAKEFORGE_STORAGE_OM_SERVICE", "seaweedfs"),
-            storage_display_name=environ.get("OPENLAKEFORGE_STORAGE_DISPLAY_NAME", "SeaweedFS S3"),
-            storage_endpoint=environ.get("OPENLAKEFORGE_STORAGE_ENDPOINT", "http://seaweedfs-s3:8333"),
-            storage_region=environ.get("OPENLAKEFORGE_STORAGE_REGION", "us-east-1"),
-            storage_bronze_bucket=environ.get(
-                "OPENLAKEFORGE_STORAGE_BRONZE_BUCKET",
-                environ.get("OPENLAKEFORGE_STORAGE_BUCKET", "lakehouse-bronze"),
-            ),
-            storage_silver_bucket=environ.get("OPENLAKEFORGE_STORAGE_SILVER_BUCKET", "lakehouse-silver"),
-            storage_gold_bucket=environ.get("OPENLAKEFORGE_STORAGE_GOLD_BUCKET", "lakehouse-gold"),
-        )
-
-
-def _parse_json_env(name: str, raw: str) -> dict:
-    try:
-        value = json.loads(raw or "{}")
-    except json.JSONDecodeError as exc:
-        raise OpenMetadataError(f"Environment variable {name} must be valid JSON: {exc}") from exc
-    if not isinstance(value, dict):
-        raise OpenMetadataError(f"Environment variable {name} must contain a JSON object.")
-    return value
-
-
-def resolve_metadata_descriptor_paths(
-    metadata_root: Path, metadata_source_dir: str
-) -> tuple[list[Path], Path, bool]:
-    """Resolve the descriptors an OpenMetadata deploy will actually read.
-
-    ``metadata_source_dir`` may name a single ``domain.yaml`` file, a single
-    domain's directory, or a multi-domain root — the same three shapes
-    :meth:`OpenMetadataDeployer.domain_files` accepts. Returns
-    ``(descriptor_paths, source_label, require_directory_match)`` so callers
-    computing the schema-FQN defaults use exactly the descriptors that will
-    be deployed, not ``metadata_root`` unconditionally. The single-file and
-    single-domain-directory shapes name an arbitrary parent directory (for
-    example a mounted ``/metadata/domain.yaml``), so ``require_directory_match``
-    is ``False`` for those.
-    """
-    if metadata_source_dir:
-        path = Path(metadata_source_dir)
-        if not path.exists():
-            raise OpenMetadataError(f"OpenMetadata metadata source does not exist: {path}")
-        if path.is_file():
-            return [path], path, False
-        if (path / "domain.yaml").is_file():
-            return [path / "domain.yaml"], path, False
-        return sorted(path.glob("*/domain.yaml")), path, True
-
-    if not metadata_root.exists():
-        raise OpenMetadataError(f"OpenMetadata metadata root does not exist: {metadata_root}")
-    return sorted(p for p in metadata_root.glob("*/domain.yaml") if p.is_file()), metadata_root, True
-
-
-def _default_schema_fqns(
-    metadata_root: str, metadata_source_dir: str, catalog_database_fqn: str, layer: str
-) -> dict[str, str]:
-    """Derive the default schema FQN contract from the domain inventory.
-
-    Used for direct local CLI execution, where the contract environment has
-    not resolved OPENLAKEFORGE_CATALOG_{SILVER,GOLD}_SCHEMA_FQNS_JSON. Reads
-    the same descriptors resolve_metadata_descriptor_paths() will actually
-    deploy, so an OPENMETADATA_METADATA_SOURCE_DIR override that names a
-    different product set still gets matching defaults.
-    """
-    descriptor_paths, source_label, require_directory_match = resolve_metadata_descriptor_paths(
-        Path(metadata_root), metadata_source_dir
-    )
-    inventory = load_domain_inventory_from_descriptors(
-        descriptor_paths, source_label=source_label, require_directory_match=require_directory_match
-    )
-    physical = inventory.resolve_physical_names(
-        catalog_database_fqn=catalog_database_fqn,
-        silver_bucket="",
-        gold_bucket="",
-        manifest_base_uri="",
-    )
-    return physical.silver_schema_fqns if layer == "silver" else physical.gold_schema_fqns
-
-
-def display_name_from_name(name: str) -> str:
-    return " ".join(part.capitalize() for part in name.replace("-", " ").replace("_", " ").split())
-
-
-def domain_description(domain: dict) -> str:
-    parts = []
-    if domain.get("description"):
-        parts.append(str(domain["description"]))
-    if domain.get("status"):
-        parts.append(f"Status: {domain['status']}")
-    medallion = domain.get("medallion")
-    if isinstance(medallion, dict) and medallion:
-        parts.append("Medallion layers:")
-        for layer, config in medallion.items():
-            if isinstance(config, dict):
-                owner = config.get("owner", "unknown")
-                description = config.get("description", "")
-                parts.append(f"- {layer}: {description} Owner: {owner}.")
-    return "\n".join(parts)
-
-
-def domain_payload(domain: dict) -> dict:
-    name = domain.get("name")
-    if not name:
-        raise OpenMetadataError("OpenMetadata domain metadata is missing required field 'name'.")
-    payload = {
-        "name": name,
-        "displayName": domain.get("displayName") or domain.get("display_name") or display_name_from_name(name),
-        "domainType": domain.get("domainType") or domain.get("domain_type") or "Source-aligned",
-        "description": domain_description(domain),
-    }
-    for optional in ["owners", "experts", "reviewers", "tags", "style", "extension"]:
-        if domain.get(optional):
-            payload[optional] = domain[optional]
-    return payload
-
-
-def product_payload(product: dict) -> dict:
-    name = product.get("name")
-    if not name:
-        raise OpenMetadataError("OpenMetadata data-product metadata is missing required field 'name'.")
-    domains = product.get("domains")
-    if domains is None and product.get("domain"):
-        domains = [product["domain"]]
-    if not domains:
-        raise OpenMetadataError(f"OpenMetadata data product '{name}' must define 'domain' or 'domains'.")
-    payload = {
-        "name": name,
-        "displayName": product.get("displayName") or product.get("display_name") or display_name_from_name(name),
-        "description": product.get("description") or "",
-        "domains": domains,
-    }
-    for optional in ["owners", "experts", "reviewers", "tags", "style", "extension"]:
-        if product.get(optional):
-            payload[optional] = product[optional]
-    return payload
-
-
-def product_entries(domain: dict) -> Iterator[dict]:
-    products = domain.get("data_products") or []
-    if not isinstance(products, list):
-        raise OpenMetadataError(f"Domain '{domain.get('name', '<unknown>')}' data_products must be a list.")
-    for product in products:
-        if not isinstance(product, dict):
-            raise OpenMetadataError(f"Unsupported data product entry in domain '{domain.get('name')}': {product!r}")
-        if not product.get("name"):
-            product["name"] = f"{domain['name']}_{product.get('id', '')}".rstrip("_")
-        if not product.get("domain") and not product.get("domains"):
-            product["domain"] = domain["name"]
-        yield product
-
-
-def product_contract_key(product: dict) -> str:
-    return product.get("asset_prefix") or product.get("name") or product.get("id") or ""
-
-
-class OpenMetadataDeployer:
-    """Stateful seeding driver; mirrors the original module functions."""
-
-    def __init__(self, config: OpenMetadataConfig, client: OpenMetadataClient):
-        self.config = config
-        self.client = client
-        self._ensured_schema_fqns: set[str] = set()
-
-    # --- product/table spec helpers ---------------------------------------
-
-    def schema_fqn_for_product(self, product: dict, table_group_key: str) -> str | None:
-        product_key = product_contract_key(product)
-        if table_group_key == "silver_tables":
-            return self.config.catalog_silver_schema_fqns.get(product_key)
-        if table_group_key == "gold_tables":
-            return self.config.catalog_gold_schema_fqns.get(product_key)
-        return None
-
-    def product_table_specs(self, product: dict):
-        for key in ["silver_tables", "gold_tables"]:
-            spec = product.get(key)
-            if not spec:
-                continue
-            schema_fqn = self.schema_fqn_for_product(product, key)
-            if not schema_fqn:
-                raise OpenMetadataError(
-                    f"Data product '{product_contract_key(product)}' table group '{key}' is not covered "
-                    "by the provider contract schema FQNs."
-                )
-            for table in spec.get("tables", []):
-                yield schema_fqn, table
-
-    def validate_deployment_inputs(self, domain_specs: list[tuple[Path, dict]]) -> None:
-        """Resolve every declared table and logical asset before metadata writes."""
-        for _, domain in domain_specs:
-            for product in product_entries(domain):
-                list(self.product_asset_entries(product))
-        self.validate_bronze_entries(domain_specs)
-
-    def provider_asset_fqn(self, product: dict, fqn):
-        if not fqn:
-            return fqn
-        table_name = fqn.rsplit(".", 1)[-1]
-        for schema_fqn, table in self.product_table_specs(product):
-            if table.get("name") == table_name:
-                return f"{schema_fqn}.{table_name}"
-        return fqn
-
-    def logical_asset_fqn(self, product: dict, asset: dict) -> str:
-        """Resolve a provider-neutral table name through the contract schemas."""
-        name = asset.get("name")
-        if not name:
-            raise OpenMetadataError(f"OpenMetadata table asset is missing 'name' or 'fqn': {asset!r}")
-        matches = [
-            f"{schema_fqn}.{name}"
-            for schema_fqn, table in self.product_table_specs(product)
-            if table.get("name") == name
-        ]
-        if not matches:
-            raise OpenMetadataError(
-                f"OpenMetadata logical table asset '{name}' is not declared in the product table contract."
-            )
-        if len(matches) > 1:
-            raise OpenMetadataError(f"OpenMetadata logical table asset '{name}' is ambiguous: {matches!r}")
-        return matches[0]
-
-    def asset_with_provider_fqn(self, product: dict, asset):
-        if isinstance(asset, str):
-            return self.provider_asset_fqn(product, asset)
-        if isinstance(asset, dict):
-            rewritten = dict(asset)
-            fqn = rewritten.get("fqn") or rewritten.get("fullyQualifiedName")
-            if not fqn and rewritten.get("name"):
-                fqn = self.logical_asset_fqn(product, rewritten)
-            if fqn:
-                rewritten["fqn"] = self.provider_asset_fqn(product, fqn)
-                rewritten.pop("fullyQualifiedName", None)
-            return rewritten
-        return asset
-
-    def product_asset_entries(self, product: dict):
-        seen = set()
-        for schema_fqn, table in self.product_table_specs(product):
-            fqn = f"{schema_fqn}.{table['name']}"
-            if fqn in seen:
-                continue
-            seen.add(fqn)
-            yield {"type": "table", "fqn": fqn}
-
-        for asset in product.get("assets", []):
-            if isinstance(asset, str):
-                fqn = asset
-            elif isinstance(asset, dict):
-                fqn = asset.get("fqn") or asset.get("fullyQualifiedName")
-            else:
-                fqn = None
-            resolved = self.asset_with_provider_fqn(product, asset)
-            if isinstance(resolved, dict):
-                fqn = resolved.get("fqn")
-            else:
-                fqn = self.provider_asset_fqn(product, fqn)
-            if fqn and fqn in seen:
-                continue
-            if fqn:
-                seen.add(fqn)
-            yield resolved
-
-    def storage_bucket_specs(self):
-        specs = [
-            (self.config.storage_bronze_bucket, "Bronze landing bucket for raw immutable source files."),
-            (
-                self.config.storage_silver_bucket,
-                "Silver bucket for Floe-validated Iceberg tables and validation reports.",
-            ),
-            (self.config.storage_gold_bucket, "Gold bucket for dbt-owned business marts."),
-        ]
-        seen = set()
-        for name, description in specs:
-            if not name or name in seen:
-                continue
-            seen.add(name)
-            yield {"name": name, "path": f"s3://{name}", "description": description}
-
-    def bronze_container_specs(self, domain_specs: list[tuple[Path, dict]]):
-        """Build provider-specific Bronze container hierarchies from domain contracts."""
-        bucket = self.config.storage_bronze_bucket
-        root_fqn = f"{self.config.storage_service}.{bucket}"
-        seen = set()
-
-        for _, domain in domain_specs:
-            for product in product_entries(domain):
-                for source in product.get("bronze") or []:
-                    source_path = source["path"]
-                    parsed = urllib.parse.urlparse(source_path)
-                    if parsed.scheme != "s3":
-                        raise OpenMetadataError(
-                            f"Data product '{product['name']}' Bronze path must use s3://: {source_path}"
-                        )
-
-                    segments = [segment for segment in parsed.path.split("/") if segment]
-                    parent_fqn = root_fqn
-                    full_path = f"s3://{bucket}"
-                    for index, segment in enumerate(segments):
-                        full_path = f"{full_path}/{segment}"
-                        current_fqn = f"{parent_fqn}.{segment}"
-                        if full_path not in seen:
-                            is_source = index == len(segments) - 1
-                            description = (
-                                source.get("description", "") if is_source else f"Bronze source prefix for {full_path}."
-                            )
-                            yield {
-                                "name": segment,
-                                "parent_fqn": parent_fqn,
-                                "path": full_path,
-                                "description": description,
-                            }
-                            seen.add(full_path)
-                        parent_fqn = current_fqn
-
-    # --- source discovery -------------------------------------------------
-
-    def domain_files(self):
-        descriptor_paths, _source_label, _require_directory_match = resolve_metadata_descriptor_paths(
-            self.config.metadata_root, self.config.metadata_source_dir
-        )
-        return descriptor_paths
-
-    # --- REST operations --------------------------------------------------
-
-    def login(self) -> None:
-        self.client.login(self.config.admin_email, self.config.admin_password)
-
-    def wait_for_openmetadata(self) -> None:
-        last_error = None
-        for _ in range(120):
-            try:
-                self.client.request("GET", "/api/v1/system/config/jwks")
-                return
-            except OpenMetadataError as exc:
-                last_error = exc
-                time.sleep(2)
-        raise OpenMetadataError(f"OpenMetadata did not become reachable: {last_error}")
-
-    def resolve_table_asset(self, asset):
-        if isinstance(asset, str):
-            asset_type = "table"
-            fqn = asset
-        elif isinstance(asset, dict):
-            asset_type = asset.get("type", "table")
-            fqn = asset.get("fqn") or asset.get("fullyQualifiedName")
-        else:
-            raise OpenMetadataError(f"Unsupported asset entry: {asset!r}")
-
-        if asset_type != "table":
-            raise OpenMetadataError(
-                f"Unsupported OpenMetadata data-product asset type '{asset_type}'. Only 'table' is supported."
-            )
-        if not fqn:
-            raise OpenMetadataError(f"OpenMetadata data-product asset is missing 'fqn': {asset!r}")
-
-        encoded_fqn = urllib.parse.quote(fqn, safe="")
-        try:
-            table = self.client.request("GET", f"/api/v1/tables/name/{encoded_fqn}?fields=domains")
-        except OpenMetadataError as exc:
-            if "HTTP 404" in str(exc):
-                return None, fqn
-            raise
-
-        table_id = table.get("id")
-        if not table_id:
-            raise OpenMetadataError(f"OpenMetadata table lookup for '{fqn}' did not return an id: {table}")
-        return {
-            "id": table_id,
-            "type": "table",
-            "name": table.get("name"),
-            "fullyQualifiedName": table.get("fullyQualifiedName", fqn),
-            "displayName": table.get("displayName"),
-            "domains": table.get("domains") or [],
-        }, None
-
-    def resolve_domain_ref(self, domain_name):
-        encoded_name = urllib.parse.quote(domain_name, safe="")
-        domain = self.client.request("GET", f"/api/v1/domains/name/{encoded_name}")
-        domain_id = domain.get("id")
-        if not domain_id:
-            raise OpenMetadataError(f"OpenMetadata domain lookup for '{domain_name}' did not return an id: {domain}")
-        return {
-            "id": domain_id,
-            "type": "domain",
-            "name": domain.get("name", domain_name),
-            "fullyQualifiedName": domain.get("fullyQualifiedName", domain_name),
-            "displayName": domain.get("displayName"),
-        }
-
-    @staticmethod
-    def _domain_matches(existing, expected) -> bool:
-        return (
-            existing.get("id") == expected.get("id")
-            or existing.get("fullyQualifiedName") == expected.get("fullyQualifiedName")
-            or existing.get("name") == expected.get("name")
-        )
-
-    def ensure_table_domains(self, table_refs, domain_refs) -> None:
-        for table_ref in table_refs:
-            existing_domains = table_ref.get("domains") or []
-            missing_domains = [
-                domain_ref
-                for domain_ref in domain_refs
-                if not any(self._domain_matches(existing, domain_ref) for existing in existing_domains)
-            ]
-            if not missing_domains:
-                continue
-            domains = existing_domains + missing_domains
-            self.client.request(
-                "PATCH",
-                f"/api/v1/tables/{table_ref['id']}",
-                payload=[{"op": "add", "path": "/domains", "value": domains}],
-                content_type="application/json-patch+json",
-            )
-            table_ref["domains"] = domains
-            print(
-                "Assigned OpenMetadata domain(s) "
-                f"{', '.join(domain['fullyQualifiedName'] for domain in missing_domains)} "
-                f"to table: {table_ref['fullyQualifiedName']}"
-            )
-
-    @staticmethod
-    def _data_product_asset_ref(table_ref):
-        return {
-            "id": table_ref["id"],
-            "type": table_ref["type"],
-            "name": table_ref.get("name"),
-            "fullyQualifiedName": table_ref.get("fullyQualifiedName"),
-            "displayName": table_ref.get("displayName"),
-        }
-
-    def ensure_storage_service(self) -> None:
-        aws_config = {"awsRegion": self.config.storage_region}
-        if self.config.storage_endpoint:
-            aws_config["endPointURL"] = self.config.storage_endpoint
-        payload = {
-            "name": self.config.storage_service,
-            "displayName": self.config.storage_display_name,
-            "serviceType": "S3",
-            "connection": {"config": {"type": "S3", "awsConfig": aws_config}},
-        }
-        self.client.request("PUT", "/api/v1/services/storageServices", payload=payload, ok_statuses=(200, 201))
-        print(f"Upserted OpenMetadata storage service: {self.config.storage_service}")
-
-    def ensure_container(self, name, parent_fqn, full_path, description) -> None:
-        payload = {
-            "name": name,
-            "service": self.config.storage_service,
-            "fullPath": full_path,
-            "description": description,
-        }
-        if parent_fqn:
-            encoded = urllib.parse.quote(parent_fqn, safe="")
-            parent = self.client.request("GET", f"/api/v1/containers/name/{encoded}")
-            parent_id = parent.get("id")
-            if not parent_id:
-                raise OpenMetadataError(
-                    f"OpenMetadata container lookup for '{parent_fqn}' did not return an id: {parent}"
-                )
-            payload["parent"] = {"id": parent_id, "type": "container"}
-        self.client.request("PUT", "/api/v1/containers", payload=payload, ok_statuses=(200, 201))
-        print(f"Upserted OpenMetadata container: {full_path}")
-
-    def ensure_database_schema(self, schema_fqn: str) -> None:
-        """Create the databaseSchema a table stub is about to reference.
-
-        Phase 1 used to pre-create these alongside the Polaris namespaces it
-        owned. Now that namespaces are reconciled in Phase 2 (ADR 0022), the
-        schema entity has to be created here instead -- OpenMetadata rejects a
-        table whose `databaseSchema` does not resolve, and the Polaris crawler
-        that would otherwise discover it has not run yet at seeding time.
-        """
-        database_fqn, _, name = schema_fqn.rpartition(".")
-        if not database_fqn or not name:
-            raise OpenMetadataError(f"Malformed schema FQN {schema_fqn!r}: expected '<service>.<database>.<schema>'")
-        if schema_fqn in self._ensured_schema_fqns:
-            return
-        self.client.request(
-            "PUT",
-            "/api/v1/databaseSchemas",
-            payload={"name": name, "database": database_fqn},
-            ok_statuses=(200, 201),
-        )
-        self._ensured_schema_fqns.add(schema_fqn)
-        print(f"Upserted OpenMetadata database schema: {schema_fqn}")
-
-    def ensure_table_stub(self, schema_fqn, name, description) -> None:
-        payload = {"name": name, "databaseSchema": schema_fqn, "columns": []}
-        if description:
-            payload["description"] = description
-        self.client.request("PUT", "/api/v1/tables", payload=payload, ok_statuses=(200, 201))
-        print(f"Upserted OpenMetadata table stub: {schema_fqn}.{name}")
-
-    def validate_bronze_entries(self, domain_specs) -> None:
-        for _, domain in domain_specs:
-            for product in product_entries(domain):
-                bronze_entries = product.get("bronze")
-                if bronze_entries is None:
-                    continue
-                if not isinstance(bronze_entries, list):
-                    raise OpenMetadataError(
-                        f"Data product '{product['name']}' Bronze entries must be an array."
-                    )
-                for index, container in enumerate(bronze_entries):
-                    if not isinstance(container, dict):
-                        raise OpenMetadataError(
-                            f"Data product '{product['name']}' Bronze entry at index {index} must be an object."
-                        )
-                    if not isinstance(container.get("name"), str) or not container["name"]:
-                        raise OpenMetadataError(
-                            f"Data product '{product['name']}' Bronze entry at index {index} "
-                            "is missing required 'name'."
-                        )
-                    if not isinstance(container.get("path"), str) or not container["path"]:
-                        raise OpenMetadataError(
-                            f"Data product '{product['name']}' Bronze entry at index {index} "
-                            "is missing required 'path'."
-                        )
-
-    def cleanup_legacy_default_database(self) -> None:
-        if not self.config.cleanup_legacy_default_database or self.config.catalog_database == "default":
-            return
-        target_fqn = self.config.catalog_database_fqn
-        legacy_fqn = f"{self.config.catalog_service}.default"
-        encoded_target = urllib.parse.quote(target_fqn, safe="")
-        encoded_legacy = urllib.parse.quote(legacy_fqn, safe="")
-
-        try:
-            self.client.request("GET", f"/api/v1/databases/name/{encoded_target}")
-        except OpenMetadataError as exc:
-            if "HTTP 404" in str(exc):
-                import sys
-
-                print(
-                    "WARN: Skipping legacy OpenMetadata database cleanup because "
-                    f"target database is missing: {target_fqn}",
-                    file=sys.stderr,
-                )
-                return
-            raise
-
-        try:
-            legacy = self.client.request("GET", f"/api/v1/databases/name/{encoded_legacy}")
-        except OpenMetadataError as exc:
-            if "HTTP 404" in str(exc):
-                return
-            raise
-
-        legacy_id = legacy.get("id")
-        if not legacy_id:
-            raise OpenMetadataError(f"OpenMetadata database lookup for '{legacy_fqn}' did not return an id: {legacy}")
-        self.client.request(
-            "DELETE",
-            f"/api/v1/databases/{legacy_id}?recursive=true&hardDelete=true",
-            ok_statuses=(200, 202, 204),
-        )
-        print(f"Deleted legacy OpenMetadata database metadata: {legacy_fqn}")
-
-    def deploy(self) -> None:
-        self.wait_for_openmetadata()
-        self.login()
-
-        domain_specs = [(path, load_domain_descriptor(path)) for path in self.domain_files()]
-        if not domain_specs:
-            raise OpenMetadataError(
-                f"No OpenMetadata domain metadata files found under {self.config.metadata_root}/<domain>/domain.yaml"
-            )
-        self.validate_deployment_inputs(domain_specs)
-
-        # Phase A+B: Object Store service and medallion bucket containers.
-        self.ensure_storage_service()
-        for container in self.storage_bucket_specs():
-            self.ensure_container(container["name"], None, container["path"], container["description"])
-        for container in self.bronze_container_specs(domain_specs):
-            self.ensure_container(
-                container["name"],
-                container["parent_fqn"],
-                container["path"],
-                container["description"],
-            )
-
-        # Phase C: Pre-seed Iceberg table stubs before the Polaris crawler runs.
-        for _, domain in domain_specs:
-            for product in product_entries(domain):
-                for schema_fqn, table in self.product_table_specs(product):
-                    self.ensure_database_schema(schema_fqn)
-                    self.ensure_table_stub(schema_fqn, table["name"], table.get("description", ""))
-        self.cleanup_legacy_default_database()
-
-        # Phase D: Upsert domains and data products from governance YAML.
-        missing_assets = []
-        for _, domain in domain_specs:
-            domain_body = domain_payload(domain)
-            self.client.request("PUT", "/api/v1/domains", payload=domain_body, ok_statuses=(200, 201))
-            print(f"Upserted OpenMetadata domain: {domain_body['name']}")
-
-            for product in product_entries(domain):
-                product_body = product_payload(product)
-                self.client.request("PUT", "/api/v1/dataProducts", payload=product_body, ok_statuses=(200, 201))
-                print(f"Upserted OpenMetadata data product: {product_body['name']}")
-                domain_refs = [self.resolve_domain_ref(domain_name) for domain_name in product_body["domains"]]
-
-                refs = []
-                for asset in self.product_asset_entries(product):
-                    ref, missing_fqn = self.resolve_table_asset(asset)
-                    if missing_fqn:
-                        missing_assets.append(missing_fqn)
-                    else:
-                        refs.append(ref)
-
-                if refs:
-                    self.ensure_table_domains(refs, domain_refs)
-                    product_name = urllib.parse.quote(product_body["name"], safe="")
-                    self.client.request(
-                        "PUT",
-                        f"/api/v1/dataProducts/{product_name}/assets/add",
-                        payload={"assets": [self._data_product_asset_ref(ref) for ref in refs], "dryRun": False},
-                        ok_statuses=(200, 201),
-                    )
-                    print(f"Attached {len(refs)} OpenMetadata asset(s) to data product: {product_body['name']}")
-
-        if missing_assets:
-            message = "\n".join(f"  - {fqn}" for fqn in sorted(set(missing_assets)))
-            guidance = (
-                "OpenMetadata table assets are not available yet:\n"
-                f"{message}\n"
-                "Run the product ETL jobs in Dagster, wait for the catalog metadata ingestion to crawl the catalog, "
-                "then rerun 'make openmetadata-metadata-deploy'."
-            )
-            if self.config.allow_missing_assets:
-                import sys
-
-                print(f"WARN: {guidance}", file=sys.stderr)
-            else:
-                raise OpenMetadataError(guidance)
