+++ tools/olf/tests/test_openmetadata_orchestration.py
@@ -5,228 +5,6 @@ import pytest
 from olf import openmetadata as om
 
 
-def test_display_name_from_name() -> None:
-    assert om.display_name_from_name("order_revenue") == "Order Revenue"
-    assert om.display_name_from_name("supply-chain") == "Supply Chain"
-
-
-def test_domain_payload_builds_description_and_defaults() -> None:
-    payload = om.domain_payload(
-        {
-            "name": "sales",
-            "description": "Sales domain",
-            "status": "active",
-            "medallion": {"bronze": {"owner": "ingest", "description": "raw."}},
-        }
-    )
-    assert payload["name"] == "sales"
-    assert payload["displayName"] == "Sales"
-    assert payload["domainType"] == "Source-aligned"
-    assert "Status: active" in payload["description"]
-    assert "- bronze: raw. Owner: ingest." in payload["description"]
-
-
-def test_product_payload_requires_domain() -> None:
-    with pytest.raises(om.OpenMetadataError):
-        om.product_payload({"name": "order_revenue"})
-    payload = om.product_payload({"name": "order_revenue", "domain": "sales"})
-    assert payload["domains"] == ["sales"]
-
-
-def test_product_entries_defaults_name_and_domain() -> None:
-    domain = {"name": "sales", "data_products": [{"id": "orders"}]}
-    products = list(om.product_entries(domain))
-    assert products[0]["name"] == "sales_orders"
-    assert products[0]["domain"] == "sales"
-
-
-def test_config_from_environment_reads_schema_fqns() -> None:
-    environ = {
-        "OPENLAKEFORGE_CATALOG_DATABASE_FQN": "aws_glue.lakehouse_dev",
-        "OPENLAKEFORGE_CATALOG_SILVER_SCHEMA_FQNS_JSON": '{"order_revenue": "svc.db.order_revenue_silver"}',
-        "OPENLAKEFORGE_CATALOG_GOLD_SCHEMA_FQNS_JSON": "{}",
-        "OPENLAKEFORGE_STORAGE_OM_SERVICE": "aws_s3",
-        "OPENLAKEFORGE_STORAGE_DISPLAY_NAME": "AWS S3",
-        "OPENLAKEFORGE_STORAGE_ENDPOINT": "",
-        "OPENLAKEFORGE_STORAGE_REGION": "eu-west-1",
-        "OPENLAKEFORGE_STORAGE_BRONZE_BUCKET": "openlakeforge-poc-bronze",
-        "OPENLAKEFORGE_STORAGE_SILVER_BUCKET": "openlakeforge-poc-silver",
-        "OPENLAKEFORGE_STORAGE_GOLD_BUCKET": "openlakeforge-poc-gold",
-    }
-    cfg = om.OpenMetadataConfig.from_environment(
-        environ,
-        base_url="http://127.0.0.1:18585/",
-        admin_email="admin@open-metadata.org",
-        admin_password="admin",
-        metadata_root="domains",
-        metadata_source_dir="",
-        allow_missing_assets=True,
-        catalog_service="polaris",
-        catalog_database="lakehouse_dev",
-        cleanup_legacy_default_database=True,
-    )
-    assert cfg.base_url == "http://127.0.0.1:18585"
-    assert cfg.catalog_database_fqn == "aws_glue.lakehouse_dev"
-    assert cfg.catalog_silver_schema_fqns == {"order_revenue": "svc.db.order_revenue_silver"}
-    assert cfg.storage_service == "aws_s3"
-    assert cfg.storage_display_name == "AWS S3"
-    assert cfg.storage_endpoint == ""
-    assert cfg.storage_region == "eu-west-1"
-    assert cfg.storage_bronze_bucket == "openlakeforge-poc-bronze"
-    assert cfg.storage_silver_bucket == "openlakeforge-poc-silver"
-    assert cfg.storage_gold_bucket == "openlakeforge-poc-gold"
-
-
-def test_config_from_environment_defaults_seed_schema_fqns_for_direct_cli() -> None:
-    cfg = om.OpenMetadataConfig.from_environment(
-        {},
-        base_url="http://x",
-        admin_email="a",
-        admin_password="p",
-        metadata_root="domains",
-        metadata_source_dir="",
-        allow_missing_assets=False,
-        catalog_service="",
-        catalog_database="",
-        cleanup_legacy_default_database=False,
-    )
-
-    assert cfg.catalog_service == "polaris"
-    assert cfg.catalog_database == "lakehouse_dev"
-    assert cfg.catalog_silver_schema_fqns["sales_order_revenue"] == (
-        "polaris.lakehouse_dev.sales_order_revenue_silver"
-    )
-    assert cfg.catalog_gold_schema_fqns["sales_order_revenue"] == (
-        "polaris.lakehouse_dev.sales_order_revenue_gold"
-    )
-
-
-def test_config_from_environment_derives_defaults_from_metadata_source_dir_override(tmp_path: Path) -> None:
-    """metadata_source_dir must drive the FQN defaults, matching what domain_files() actually deploys.
-
-    metadata_root here does not even exist, so the defaults could only have
-    come from the override.
-    """
-    source_dir = tmp_path / "override"
-    source_dir.mkdir()
-    (source_dir / "domain.yaml").write_text(
-        """apiVersion: openlakeforge.io/v1alpha2
-kind: Domain
-name: override
-displayName: Override
-description: Override-only domain.
-status: active
-data_products:
-  - id: widgets
-    name: override_widgets
-    displayName: Override Widgets
-    description: Widgets.
-    status: active
-    asset_prefix: override_widgets
-    bronze:
-      - name: source
-        path: s3://lakehouse-bronze/override/widgets/source
-    silver_tables:
-      tables:
-        - name: source
-    gold_tables:
-      tables:
-        - name: mart_widgets
-"""
-    )
-
-    cfg = om.OpenMetadataConfig.from_environment(
-        {},
-        base_url="http://x",
-        admin_email="a",
-        admin_password="p",
-        metadata_root=str(tmp_path / "does-not-exist"),
-        metadata_source_dir=str(source_dir),
-        allow_missing_assets=False,
-        catalog_service="polaris",
-        catalog_database="lakehouse_dev",
-        cleanup_legacy_default_database=False,
-    )
-
-    assert cfg.catalog_silver_schema_fqns == {"override_widgets": "polaris.lakehouse_dev.override_widgets_silver"}
-    assert cfg.catalog_gold_schema_fqns == {"override_widgets": "polaris.lakehouse_dev.override_widgets_gold"}
-
-
-def test_config_from_environment_accepts_a_standalone_metadata_file_override(tmp_path: Path) -> None:
-    """A single-file override's parent directory name must not have to match the domain.
-
-    Mirrors a mounted or temporary path like /metadata/domain.yaml, where the
-    parent directory carries no significance — only the file's own `name:`
-    field identifies the domain.
-    """
-    metadata_dir = tmp_path / "metadata"
-    metadata_dir.mkdir()
-    descriptor_path = metadata_dir / "domain.yaml"
-    descriptor_path.write_text(
-        """apiVersion: openlakeforge.io/v1alpha2
-kind: Domain
-name: sales
-displayName: Sales
-description: Standalone sales descriptor.
-status: active
-data_products:
-  - id: orders
-    name: sales_orders
-    displayName: Sales Orders
-    description: Orders.
-    status: active
-    asset_prefix: sales_orders
-    bronze:
-      - name: source
-        path: s3://lakehouse-bronze/sales/orders/source
-    silver_tables:
-      tables:
-        - name: source
-    gold_tables:
-      tables:
-        - name: mart_orders
-"""
-    )
-
-    cfg = om.OpenMetadataConfig.from_environment(
-        {},
-        base_url="http://x",
-        admin_email="a",
-        admin_password="p",
-        metadata_root=str(tmp_path / "does-not-exist"),
-        metadata_source_dir=str(descriptor_path),
-        allow_missing_assets=False,
-        catalog_service="polaris",
-        catalog_database="lakehouse_dev",
-        cleanup_legacy_default_database=False,
-    )
-
-    assert cfg.catalog_silver_schema_fqns == {"sales_orders": "polaris.lakehouse_dev.sales_orders_silver"}
-    deployer = om.OpenMetadataDeployer(cfg, om.OpenMetadataClient(cfg.base_url))
-    assert deployer.domain_files() == [descriptor_path]
-
-
-def test_config_from_environment_preserves_explicit_empty_schema_contract() -> None:
-    cfg = om.OpenMetadataConfig.from_environment(
-        {
-            "OPENLAKEFORGE_CATALOG_SILVER_SCHEMA_FQNS_JSON": "{}",
-            "OPENLAKEFORGE_CATALOG_GOLD_SCHEMA_FQNS_JSON": "{}",
-        },
-        base_url="http://x",
-        admin_email="a",
-        admin_password="p",
-        metadata_root="domains",
-        metadata_source_dir="",
-        allow_missing_assets=False,
-        catalog_service="polaris",
-        catalog_database="lakehouse_dev",
-        cleanup_legacy_default_database=False,
-    )
-
-    assert cfg.catalog_silver_schema_fqns == {}
-    assert cfg.catalog_gold_schema_fqns == {}
-
-
 def test_storage_bucket_specs_dedup() -> None:
     cfg = om.OpenMetadataConfig.from_environment(
         {},
@@ -560,53 +338,6 @@ data_products:
         cleanup_legacy_default_database=False,
     )
     return om.OpenMetadataDeployer(cfg, om.OpenMetadataClient(cfg.base_url))
-
-
-def test_ensure_database_schema_creates_the_schema_a_table_will_reference(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    """Phase 1 stopped pre-creating databaseSchemas when Polaris namespace
-    lifecycle moved to Phase 2, so seeding has to create them itself."""
-    deployer = _single_product_deployer(tmp_path)
-    calls: list[tuple[str, dict]] = []
-
-    def request(method: str, path: str, *, payload=None, **_kwargs):
-        calls.append((path, payload or {}))
-        return {}
-
-    monkeypatch.setattr(deployer.client, "request", request)
-
-    deployer.ensure_database_schema("polaris.lakehouse_dev.sales_order_revenue_silver")
-
-    assert calls == [
-        (
-            "/api/v1/databaseSchemas",
-            {"name": "sales_order_revenue_silver", "database": "polaris.lakehouse_dev"},
-        )
-    ]
-
-
-def test_ensure_database_schema_is_created_once_per_schema(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    deployer = _single_product_deployer(tmp_path)
-    calls: list[str] = []
-
-    monkeypatch.setattr(
-        deployer.client, "request", lambda method, path, **_kwargs: calls.append(path) or {}
-    )
-
-    deployer.ensure_database_schema("polaris.lakehouse_dev.sales_order_revenue_silver")
-    deployer.ensure_database_schema("polaris.lakehouse_dev.sales_order_revenue_silver")
-
-    assert calls == ["/api/v1/databaseSchemas"]
-
-
-def test_ensure_database_schema_rejects_a_malformed_fqn(tmp_path: Path) -> None:
-    deployer = _single_product_deployer(tmp_path)
-
-    with pytest.raises(om.OpenMetadataError, match="Malformed schema FQN"):
-        deployer.ensure_database_schema("sales_order_revenue_silver")
 
 
 def test_deploy_creates_each_schema_before_seeding_its_tables(
